surveillance/requirements.txt
# FILE: requirements.txt

ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.24.0
scipy>=1.10.0
scikit-image>=0.20.0
pandas>=2.0.0
pyyaml>=6.0
tqdm>=4.65.0
matplotlib>=3.7.0
streamlit>=1.28.0
